[Refactor] FavorDatePickerTextField에 `.custom` 타입을 추가하여 범용적으로 사용할 수 있도록 리팩토링
